+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Changed [Sponge-Property] Enable multiple properties.
 - Added [Watch-Process][] function
 - Refactor [gyo][] function